requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Storage
minio==7.2.0

# Task Queue
celery==5.3.4
redis==5.0.1

# Data Processing
pandas==2.1.3
polars==0.19.19
pyarrow==14.0.1
duckdb==0.9.2
openpyxl==3.1.2
python-docx==1.1.0
PyPDF2==3.0.1

# ML/AI
scikit-learn==1.3.2
sentence-transformers==2.2.2
great-expectations==0.18.7
openai==1.3.7
anthropic==0.7.7

# Scraping
scrapy==2.11.0
playwright==1.40.0
beautifulsoup4==4.12.2
requests==2.31.0

# Blockchain
web3==6.11.3
solana==0.30.2

# Utils
python-dotenv==1.0.0
loguru==0.7.2
httpx==0.25.2
python-magic==0.4.27
scipy==1.11.4
